modify execute file

+++ utils/layers.py
@@ -73,7 +73,7 @@ def sp_attn_head(seq, out_sz, adj_mat, adj_2, activation, nb_nodes,W1, W2, in_dr
                 values=tf.nn.leaky_relu(logits.values), #对value进行激活
                 dense_shape=logits.dense_shape) #SparseTensor(indices=Tensor(shape(?, 2)), values=Tensor(shape(?,)), dense_shape=Tensor(shape(2,)))
         adj_2 = adj_2.tocoo()
-        adj_2 = tf.SparseTensor(np.vstack((adj_2.col, adj_2.row)).transpose(), adj_2.data, adj_2.shape)
+        adj_2 = tf.SparseTensor(np.vstack((adj_2.col, adj_2.row)).transpose(), tf.cast(adj_2.data, tf.float32), adj_2.shape)
         lrelu = tf.sparse_add(lrelu, adj_2)
         coefs = tf.sparse_softmax(lrelu)        #SparseTensor(indices=Tensor(shape(?, 2)), values=Tensor(shape(?,)), dense_shape=Tensor(shape(2,)))
